install and register arbitrary libraries
