+++ src/configuration.ts
@@ -1,10 +1,11 @@
 export interface Configuration {
   assets?: Array<string | GlobRule>;
+  modules?: Array<ModuleRef>;
+}
 
-  modules?: Array<{
-    name: string;
-    namespace: string;
-  }>;
+export interface ModuleRef {
+  name: string;
+  namespace: string;
 }
 
 export interface GlobRule {
