+++ src/utils.ts
@@ -3,7 +3,6 @@ import path from 'path';
 import fs from 'fs';
 import prettier from 'prettier';
 
-import TsUtils from './ts-utils';
 import log from './log';
 import { Configuration } from './configuration';
 
@@ -44,31 +43,6 @@ export function install(lib: string): boolean {
   }
 
   return true;
-}
-
-export function registerModules(
-  moduleName: string,
-  config: Configuration,
-  skipFormat?: boolean
-): void {
-  if (config && config.modules && config.modules.length > 0) {
-    log.info('registering modules');
-
-    const modulePath = path.join(
-      process.cwd(),
-      'src/app',
-      `${moduleName}.module.ts`
-    );
-
-    const source = fs.readFileSync(modulePath).toString();
-    const tsUtils = new TsUtils();
-
-    let sourceFile = tsUtils.parse(source);
-    sourceFile = tsUtils.registerModules(sourceFile, config.modules);
-
-    const output = tsUtils.renderFile(sourceFile);
-    fs.writeFileSync(modulePath, skipFormat ? output : format(output));
-  }
 }
 
 export function version(): string {
